third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/external/wpt/css/css-shapes/spec-examples/shape-outside-018-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: shape-margin offset from a polygonal shape-outside Uncaught Error: assert_approx_equals: Line 0 is positioned properly expected 48 +/- 2 but got 38
Harness: the test ran to completion.
